refactor(ios): move the live view out of the simulator panel
`ChatIosSimulatorPanel` carried both live-view backends, their state, refs,
callbacks and effects, the chip, the blocker, the restart path and the pointer
mapping — about 1,200 lines that touch no launch, no inspector, no preview and
no attach-to-chat. The panel goes from 4,340 lines to 3,066 and stops knowing
that H.264 exists.

A pure move. No test changed, and the 76 panel tests that pin the parking-hold
lifecycle, the capture-token contract and the h264 retry budget pass untouched.

The orderings that are load-bearing are preserved and now documented in one
place rather than inferred from where a declaration happened to sit:

- The h264 retry effect stays declared before the main live-view effect. The
  interleaving of their cleanups and setups is what stops a stale retry from
  rebuilding a port forward to a device the drawer has moved off.
- The teardown effect stays declared after it. React runs cleanups in
  declaration order, so the live effect clears the capture token first and the
  teardown then sees null and stops the stream itself. Reversed, a start still
  in flight declines to stop the stream it brought up.
- The capture token is still armed synchronously before the async prelude, and
  the prelude's five steps keep their order.
- All seven parking-hold release sites moved together. The token narrowing is
  what makes a double release a no-op, and a release left behind would
  decrement another drawer's holder.
- `trackWindowVideoFrames`, `attachLiveStream` and `setVideoNode` keep their
  dependency arrays, so the callback ref on the video element keeps its
  identity and a playing video is never detached by a re-render.

`handleBlockerAction` reaches `launch` through the existing `launchRef` rather
than the callback, because `launch` is declared after the cluster and a hook
cannot receive something that does not exist yet. The panel already used that
ref the same way elsewhere.

The hook takes `imageRef` because `liveWidth`/`liveHeight` fall back to the
inspect image when there is no live visual and no video. That is the one
argument that reads oddly, and it is documented where it is declared.
